@@ -1,45 +1,30 @@
 {
-  "name": "nuxt",
+  "name": "nuxt-app",
   "version": "1.0.0",
-  "description": "Kiosk app frontend",
+  "description": "My astonishing Nuxt.js project",
   "author": "Seiya Kobayashi",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_ENV=development nodemon server/index.js --watch server",
+    "dev": "nuxt",
     "build": "nuxt build",
-    "start": "cross-env NODE_ENV=production node server/index.js",
+    "start": "nuxt start",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "lint-fix": "eslint --fix --ext .js,.vue --ignore-path .gitignore .",
-    "precommit": "npm run lint",
-    "test": "jest"
+    "lint-fix": "eslint --fix --ext .js,.vue --ignore-path .gitignore ."
   },
   "dependencies": {
-    "cross-env": "^5.2.0",
-    "nuxt": "^2.3.4",
-    "express": "^4.16.4",
-    "@nuxtjs/axios": "^5.3.6"
+    "nuxt": "^2.0.0"
   },
   "devDependencies": {
-    "@nuxtjs/eslint-config": "^0.0.1",
-    "@vue/test-utils": "^1.0.0-beta.27",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-eslint": "^8.2.1",
-    "babel-jest": "^23.6.0",
-    "eslint": "^5.0.1",
-    "eslint-config-google": "^0.11.0",
-    "eslint-config-standard": ">=12.0.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-import": ">=2.14.0",
-    "eslint-plugin-jest": ">=21.24.1",
-    "eslint-plugin-node": ">=7.0.1",
-    "eslint-plugin-promise": ">=4.0.1",
-    "eslint-plugin-standard": ">=4.0.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "jest": "^23.6.0",
-    "nodemon": "^1.18.9",
-    "stylus": "^0.54.5",
+    "@nuxtjs/eslint-config": "^1.0.1",
+    "@nuxtjs/eslint-module": "^1.0.0",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^6.1.0",
+    "eslint-plugin-nuxt": ">=0.4.2",
+    "path": "^0.12.7",
+    "reset-css": "^4.0.1",
+    "stylus": "^0.54.7",
     "stylus-loader": "^3.0.2",
-    "vue-jest": "^3.0.2"
+    "vue-svg-loader": "^0.12.0"
   }
 }
