@@ -1,54 +1,48 @@
-const pkg = require('./package')
+import path from 'path';
 
-module.exports = {
-  mode: 'spa',
-
+export default {
+  mode: 'universal',
   /*
   ** Headers of the page
   */
   head: {
-    title: pkg.name,
+    title: 'Activity Kiosk',
     meta: [
       { charset: 'utf-8' },
       { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-      { hid: 'description', name: 'description', content: pkg.description },
+      { hid: 'description', name: 'description', content: 'List of activities you can do in/around this park' },
     ],
     link: [
       { rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' },
     ],
   },
-
   /*
   ** Customize the progress-bar color
   */
-  loading: { color: '#fff' },
-
+  loading: false,
   /*
   ** Global CSS
   */
   css: [
+    '@/assets/styles/global.styl',
   ],
-
   /*
   ** Plugins to load before mounting the App
   */
   plugins: [
   ],
-
+  /*
+  ** Nuxt.js dev-modules
+  */
+  buildModules: [
+    // Doc: https://github.com/nuxt-community/eslint-module
+    '@nuxtjs/eslint-module',
+  ],
   /*
   ** Nuxt.js modules
   */
   modules: [
-    // Doc: https://axios.nuxtjs.org/usage
-    '@nuxtjs/axios',
   ],
-  /*
-  ** Axios module configuration
-  */
-  axios: {
-    // See https://github.com/nuxt-community/axios-module#options
-  },
-
   /*
   ** Build configuration
   */
@@ -56,7 +50,7 @@ module.exports = {
     /*
     ** You can extend webpack config here
     */
-    extend(config, ctx) {
+    extend (config, ctx) {
       // Run ESLint on save
       if (ctx.isDev && ctx.isClient) {
         config.module.rules.push({
@@ -64,8 +58,30 @@ module.exports = {
           test: /\.(js|vue)$/,
           loader: 'eslint-loader',
           exclude: /(node_modules)/,
-        })
+        });
       }
+
+      // use vue-svg-loader for .svg files
+      const svgRule = config.module.rules.find((rule) => rule.test.test('.svg'));
+      svgRule.test = /\.(png|jpe?g|gif|webp)$/;
+      config.module.rules.push({
+        test: /\.svg$/,
+        loader: 'vue-svg-loader',
+      });
+
+      // add aliases
+      config.resolve.alias['@styles'] = resolve('assets/styles');
+      config.resolve.alias['@images'] = resolve('assets/images');
+      config.resolve.alias['@fonts'] = resolve('assets/fonts');
+      config.resolve.alias['@comps'] = resolve('components');
     },
   },
+};
+
+/*
+ * resolve
+ * @return {!String} absolute path
+ */
+function resolve (relPath) {
+  return path.join(__dirname, relPath);
 }
