@@ -9,8 +9,14 @@ module.exports = {
   },
   extends: [
     '@nuxtjs',
-    'google'
+    'plugin:nuxt/recommended'
   ],
   // add your custom rules here
-  rules: {}
+  rules: {
+    'semi': ['error', 'always'],
+    'comma-dangle': ['error', 'always-multiline'],
+    'no-console': 'off',
+    'arrow-parens': ['error', 'always'],
+    'curly': 'off',
+  },
 }
